refactor(request): move request traits to Traits directory
- Move NoAuthorizeTrait and HttpMethodTrait to Traits directory
- Update namespace from Trait to Traits in multiple files
- Adjust import statements in various request files to use new namespace
